app/requirements.txt
confluent-kafka
pyarrow
boto3
pandas
python-dotenv